Refactored: Removed ConflictCard component
